# openapi.yaml
openapi: 3.0.3
info:
  title: BCR CH 6
  version: 1.0.0
  contact: {}
servers:
  - url: 127.0.0.1
  - url: "{{base_url}}"
paths:
...
  /users/:
    get:
      tags:
        - Users
      summary: Get All Users
      description: Get All Users
      operationId: getAllUsers
      responses:
        "200":
          description: Get All Users
          headers:
            Content-Length:
              schema:
                type: string
                example: "1007"
            Date:
              schema:
                type: string
                example: Mon, 03 Jun 2024 11:32:14 GMT
            ETag:
              schema:
                type: string
                example: W/"3ef-Q64ff/flKC4aLuwNhlIVCLDkJvc"
            X-Powered-By:
              schema:
                type: string
                example: Express
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: success
                  users:
                    type: array
                    items:
                      type: object
                      properties:
                        created_at:
                          type: string
                          example: 2024-06-02T17:56:10.606Z
                        deleted_at:
                          nullable: true
                          example: null
                        email:
                          type: string
                          nullable: true
                          example: suadmin@bcr.id
                        id:
                          type: number
                          example: 1
                        name:
                          type: string
                          example: superadmin
                        password:
                          type: string
                          example: $2b$10$ET7ZZy07UAiSW2FEONrUMO/P8FMDtGJ8e5Lj6BEDPUkcW0ex2YORC
                        role:
                          type: string
                          example: superadmin
                        updated_at:
                          type: string
                          nullable: true
                          example: null
                        username:
                          type: string
                          example: superadmin
                    example:
                      - created_at: 2024-06-02T17:56:10.606Z
                        deleted_at: null
                        email: suadmin@bcr.id
                        id: 1
                        name: superadmin
                        password: $2b$10$ET7ZZy07UAiSW2FEONrUMO/P8FMDtGJ8e5Lj6BEDPUkcW0ex2YORC
                        role: superadmin
                        updated_at: null
                        username: superadmin
                      - created_at: 2024-06-03T11:15:03.840Z
                        deleted_at: null
                        email: null
                        id: 2
                        name: admin agak biasa
                        password: $2b$10$kCb/BGp4jw3tXsp0btSMfOCFzal8XQ8oZcU5AueW5bKbLtjZxHtDm
                        role: admin
                        updated_at: 2024-06-03T11:19:46.599Z
                        username: admin@bcr.id
                      - created_at: 2024-06-03T11:20:39.679Z
                        deleted_at: null
                        email: null
                        id: 4
                        name: orang biasa
                        password: $2b$10$Zs9wHPi8nPIOTCHjLPqOH.zGVVaabqI0U7fq2Htc6F6LGBIsmxq8m
                        role: user
                        updated_at: null
                        username: orangbiasa
                      - created_at: 2024-06-03T11:29:00.161Z
                        deleted_at: null
                        email: null
                        id: 5
                        name: orang biasa
                        password: $2b$10$t1U9crWniYJbxVvSua.EQOjgaV1Hx704dlRstIoWX3PxL8cW0qBLO
                        role: user
                        updated_at: null
                        username: orangtidakbiasa
              examples:
                Get All Users:
                  value:
                    message: success
                    users:
                      - created_at: 2024-06-02T17:56:10.606Z
                        deleted_at: null
                        email: suadmin@bcr.id
                        id: 1
                        name: superadmin
                        password: $2b$10$ET7ZZy07UAiSW2FEONrUMO/P8FMDtGJ8e5Lj6BEDPUkcW0ex2YORC
                        role: superadmin
                        updated_at: null
                        username: superadmin
                      - created_at: 2024-06-03T11:15:03.840Z
                        deleted_at: null
                        email: null
                        id: 2
                        name: admin agak biasa
                        password: $2b$10$kCb/BGp4jw3tXsp0btSMfOCFzal8XQ8oZcU5AueW5bKbLtjZxHtDm
                        role: admin
                        updated_at: 2024-06-03T11:19:46.599Z
                        username: admin@bcr.id
                      - created_at: 2024-06-03T11:20:39.679Z
                        deleted_at: null
                        email: null
                        id: 4
                        name: orang biasa
                        password: $2b$10$Zs9wHPi8nPIOTCHjLPqOH.zGVVaabqI0U7fq2Htc6F6LGBIsmxq8m
                        role: user
                        updated_at: null
                        username: orangbiasa
                      - created_at: 2024-06-03T11:29:00.161Z
                        deleted_at: null
                        email: null
                        id: 5
                        name: orang biasa
                        password: $2b$10$t1U9crWniYJbxVvSua.EQOjgaV1Hx704dlRstIoWX3PxL8cW0qBLO
                        role: user
                        updated_at: null
                        username: orangtidakbiasa
      security:
        - bearerAuth: []
    post:
      tags:
        - Users
      summary: Add new user
      description: Add new user
      operationId: addNewUser
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  example: admin biasa
                password:
                  type: string
                  example: admin9632
                role:
                  type: string
                  example: admin
                username:
                  type: string
                  example: admin@bcr.id
            examples:
              Add new user:
                value:
                  name: admin biasa
                  password: admin9632
                  role: admin
                  username: admin@bcr.id
      responses:
        "200":
          description: ""
      security:
        - bearerAuth: []
  /users/{id}:
    get:
      tags:
        - Users
      summary: Get User by Id
      description: Get User by Id
      operationId: getUserById
      responses:
        "200":
          description: User Id 2
          headers:
            Content-Length:
              schema:
                type: string
                example: "289"
            Date:
              schema:
                type: string
                example: Mon, 03 Jun 2024 11:32:20 GMT
            ETag:
              schema:
                type: string
                example: W/"121-0eguHZYhoxw+67veaaSwZyvJH3c"
            X-Powered-By:
              schema:
                type: string
                example: Express
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: success
                  user:
                    type: object
                    properties:
                      created_at:
                        type: string
                        example: 2024-06-03T11:15:03.840Z
                      deleted_at:
                        nullable: true
                        example: null
                      email:
                        nullable: true
                        example: null
                      id:
                        type: number
                        example: 2
                      name:
                        type: string
                        example: admin agak biasa
                      password:
                        type: string
                        example: $2b$10$kCb/BGp4jw3tXsp0btSMfOCFzal8XQ8oZcU5AueW5bKbLtjZxHtDm
                      role:
                        type: string
                        example: admin
                      updated_at:
                        type: string
                        example: 2024-06-03T11:19:46.599Z
                      username:
                        type: string
                        example: admin@bcr.id
              examples:
                User Id 2:
                  value:
                    message: success
                    user:
                      created_at: 2024-06-03T11:15:03.840Z
                      deleted_at: null
                      email: null
                      id: 2
                      name: admin agak biasa
                      password: $2b$10$kCb/BGp4jw3tXsp0btSMfOCFzal8XQ8oZcU5AueW5bKbLtjZxHtDm
                      role: admin
                      updated_at: 2024-06-03T11:19:46.599Z
                      username: admin@bcr.id
      security:
        - bearerAuth: []
    put:
      tags:
        - Users
      summary: Update User by Id
      description: Update User by Id
      operationId: updateUserById
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  example: admin agak biasa
            examples:
              Update User by Id:
                value:
                  name: admin agak biasa
      responses:
        "200":
          description: ""
      security:
        - bearerAuth: []
    delete:
      tags:
        - Users
      summary: Delete User by Id
      description: Delete User by Id
      operationId: deleteUserById
      responses:
        "200":
          description: Delete User by Id
          headers:
            Content-Length:
              schema:
                type: string
                example: "43"
            Date:
              schema:
                type: string
                example: Mon, 03 Jun 2024 11:32:45 GMT
            ETag:
              schema:
                type: string
                example: W/"2b-eYU+EJxwocl62DYzb4fnvPS6Y2o"
            X-Powered-By:
              schema:
                type: string
                example: Express
          content:
            application/json:
              schema:
                type: object
                properties:
                  isDeleted:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: User deleted
              examples:
                Delete User by Id:
                  value:
                    isDeleted: true
                    message: User deleted
      security:
        - bearerAuth: []
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          example: "3"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
...